feat: add comments to explain code mods may need to edit
