+++ js/pages/Leaderboard.js
@@ -1,4 +1,5 @@
-import { fetchLeaderboard, lightPackColor, darkPackColor } from '../content.js';
+import { fetchLeaderboard } from '../content.js';
+import { lightPackColor, darkPackColor } from '../config.js';
 import { localize, rgbaBind } from '../util.js';
 import { store } from '../main.js';
